requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.2
frozenlist==1.4.1
grpcio==1.65.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isodate==0.6.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-path==0.3.3
jsonschema-spec==0.2.4
jsonschema-specifications==2023.7.1
lazy-object-proxy==1.10.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.3.0
motor==3.5.1
multidict==6.0.5
nest-asyncio==1.6.0
numpy==1.26.4
openai==0.27.10
openapi-core==0.18.2
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
packaging==24.1
parse==1.20.2
pathable==0.4.3
prance==23.6.21.0
pybars4==0.9.13
pydantic==1.10.17
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
PyMeta3==0.5.1
pymongo==4.8.0
python-dotenv==1.0.0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.12.25
requests==2.32.3
rfc3339-validator==0.1.4
rich==13.7.1
rpds-py==0.19.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scipy==1.14.0
semantic-kernel==0.3.8.dev0
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4
